requirements/base.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.8
django-import-export==2.6.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.2
MarkupPy==1.14
MarkupSafe==2.0.1
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.2
tablib==3.0.0
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0
